+++ runtime-parent/runtime-compute/src/test/java/com/speedment/runtime/compute/ToIntTest.java
@@ -75,7 +75,7 @@ public final class ToIntTest extends AbstractToTest<ToInt<String>> {
     public void testCompose() {
         strings().forEach(s -> {
             final ToIntNullable<String> composed = instance.compose(str -> str + "A");
-            assertEquals((long) mapper.applyAsLong(s + "A"), composed.applyAsInt(s));
+            assertEquals(mapper.applyAsLong(s + "A"), composed.applyAsInt(s));
         });
     }
